% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_to_nimble.R
\name{formula_to_nimble}
\alias{formula_to_nimble}
\title{Title}
\usage{
formula_to_nimble(
  formula,
  data,
  model = NULL,
  family = list(),
  delay_link = list(),
  D = NULL,
  model_window = NULL,
  aggregate = FALSE,
  nested = FALSE,
  nested.censored = NULL,
  priors = list(),
  nchains = 1
)
}
\arguments{
\item{formula}{A list with formula's for model of the response variables 'delay', if you wish to fit a 'delay' model. Additionally a formula for the 'totals' response if you wish to fit a joint model.}

\item{data}{A list with data for the delay response variable, and the totals response variable (for a 'joint' model), and any covariates.}

\item{model}{Argument set to 'delay' if only modelling the delay respnse variable, or 'joint' if you wish to jointly model the delay and totals.}

\item{family}{Set distributions for model of the delay and the totals.}

\item{delay_link}{Set the link function for the delay proportions.}

\item{D}{Set the number of delays you want to model.}

\item{model_window}{Set a window length for the nowcasting model.}

\item{aggregate}{Set to TRUE is data has been aggregated (e.g. by region or age).}

\item{nested}{Set to TRUE if you with to jointly model a nested proportion of the totals.}

\item{nested.censored}{Set the number of time steps into the past the nested count of the totals is believed to be censored.}

\item{priors}{Set priors for model parameters.}

\item{nchains}{Set number of MCMC chains.}
}
\value{
A list of nimble objects.
}
\description{
Title
}
\examples{
Takes a formula and gives nimble objects.
}
